Trainer: backend interface
